backend/requirements.txt
streamlit
pandas
numpy
plotly
python-binance
statsmodels
sqlite-utils
